+++ .circleci/jobs/package-test.yml
@@ -11,7 +11,7 @@ steps:
   - run:
       name: Test CLI command
       command: |
-        poetry run myAES test-command --repeat "test phrase"
+        poetry run my-aes test-command --repeat "test phrase"
 
         if [[ $(!!) != "test phrase" ]]; then
